+++ Tests/ConfigCache/RegisterTest.php
@@ -93,18 +93,20 @@ class RegisterTest extends RegisterTestCase
 
     public function testSetAppConfig()
     {
-        $register = $this->createRegisterMock();
-        $config   = array('aaa' => 'bbb');
+        $register        = $this->createRegisterMock();
+        $serviceRegister = $this->util->getProperty($register, 'serviceRegister');
+        $config          = array('aaa' => 'bbb');
         $register->setAppConfig($config);
-        $this->assertSame($config, $this->util->getProperty($register, 'appConfig'));
+        $this->assertSame($config, $this->util->getProperty($serviceRegister, 'appConfig'));
     }
 
     public function testSetTag()
     {
-        $register = $this->createRegisterMock();
-        $tag      = 'test_tag';
+        $register        = $this->createRegisterMock();
+        $serviceRegister = $this->util->getProperty($register, 'serviceRegister');
+        $tag             = 'test_tag';
         $register->setTag($tag);
-        $this->assertSame($tag, $this->util->getProperty($register, 'tag'));
+        $this->assertSame($tag, $this->util->getProperty($serviceRegister, 'tag'));
     }
 
     /**
@@ -592,13 +594,16 @@ class RegisterTest extends RegisterTestCase
      */
     public function testInitializeResources($resources, $expectedDirs, $expectedFiles, $expectedMethodCalls)
     {
-        $internalMethod = 'setCacheDefinition';
-        list($register, ) = $this->createRegisterMockAndContainer(array($internalMethod));
+        list($register, ) = $this->createRegisterMockAndContainer();
         $id = 'register_test';
+        // set a mock to assert method calls
+        $serviceRegister = $this->createServiceRegisterMock(array('registerConfigCache'));
+        $this->util->setProperty($register, 'serviceRegister', $serviceRegister);
+
         // only assert calling method
-        $register
+        $serviceRegister
             ->expects($this->exactly($expectedMethodCalls))
-            ->method($internalMethod)
+            ->method('registerConfigCache')
             ->willReturn(null)
             ;
         $this->util->getProperty($register, 'idBuilder')->setBundleId($id);
@@ -681,12 +686,15 @@ class RegisterTest extends RegisterTestCase
      */
     public function testInitializeAllResources($bundles, $resources, $expectedDirs, $expectedFiles, $expectedMethodCalls)
     {
-        $internalMethod = 'setCacheDefinition';
-        list($register, ) = $this->createRegisterMockAndContainer(array($internalMethod));
+        list($register, ) = $this->createRegisterMockAndContainer();
+        // set a mock to assert method calls
+        $serviceRegister = $this->createServiceRegisterMock(array('registerConfigCache'));
+        $this->util->setProperty($register, 'serviceRegister', $serviceRegister);
+
         // only assert calling method
-        $register
+        $serviceRegister
             ->expects($this->exactly($expectedMethodCalls))
-            ->method($internalMethod)
+            ->method('registerConfigCache')
             ->willReturn(null)
             ;
         // $register->initializeAllResources()
@@ -814,10 +822,13 @@ class RegisterTest extends RegisterTestCase
      */
     public function testPostInitializeResources(array $files, array $dirs, $expectedMethodCalls)
     {
-        $register = $this->createRegisterMock(array('setCacheDefinition'));
-        $register
+        $register = $this->createRegisterMock();
+        // set a mock to assert method calls
+        $serviceRegister = $this->createServiceRegisterMock(array('registerConfigCache'));
+        $this->util->setProperty($register, 'serviceRegister', $serviceRegister);
+        $serviceRegister
             ->expects($expectedMethodCalls ? $this->once() : $this->never())
-            ->method('setCacheDefinition')
+            ->method('registerConfigCache')
             ->willReturn(null)
             ;
         $this->util
@@ -1035,103 +1046,6 @@ class RegisterTest extends RegisterTestCase
     }
 
     /**
-     * test setCacheDefinition and createCacheDefinition in this method
-     *
-     * @dataProvider setCacheDefinitionProvider
-     */
-    public function testSetCacheDefinition($tag)
-    {
-        list($register, $container) = $this->createRegisterMockAndContainer();
-        $id = 'register_test';
-        $this->preSetCacheDefinition($register, $tag, $id);
-
-        // differ by setCacheDefinitionByAlias
-        $register->setConfiguration($configuration = new RegisterConfiguration());
-        // setCacheDefinition
-        $this->util->invoke($register, 'setCacheDefinition');
-
-        // Definition
-        $definition = $this->postSetCacheDefinition($container, $tag, $id);
-        // DefinitionDecorator
-        $parent     = $container->getDefinition($definition->getParent());
-
-        // assert addMethodCalls simplified
-        $actualCalls = array_merge($parent->getMethodCalls(), $definition->getMethodCalls());
-        $this->assertSame('setArrayAccess', $actualCalls[0][0]);
-        $this->assertInstanceOf('Symfony\Component\DependencyInjection\Reference', $actualCalls[0][1][0]);
-        // differ by setCacheDefinitionByAlias
-        $this->assertSame('setConfiguration', $actualCalls[1][0]);
-        $this->assertInstanceOf('Symfony\Component\DependencyInjection\Reference', $actualCalls[1][1][0]);
-
-        // assert(Configuration)
-        $configId = $this->util->getProperty($register, 'idBuilder')->buildConfigurationId($configuration);
-        $this->assertTrue($container->hasDefinition($configId));
-        $definition = $container->getDefinition($configId);
-        $this->assertFalse($definition->isPublic());
-        $this->assertSame('YahooJapan\ConfigCacheBundle\Tests\Fixtures\RegisterConfiguration', $definition->getClass());
-        $this->assertSame(0, count($definition->getArguments()));
-    }
-
-    /**
-     * @return array ($tag)
-     */
-    public function setCacheDefinitionProvider()
-    {
-        return array(
-            // no tag
-            array(null),
-            // has tag
-            array('test_tag'),
-        );
-    }
-
-    /**
-     * @dataProvider setCacheDefinitionProvider
-     */
-    public function testSetCacheDefinitionByAlias($tag)
-    {
-        list($register, $container) = $this->createRegisterMockAndContainer();
-        $id = 'register_test';
-        $this->preSetCacheDefinition($register, $tag, $id);
-
-        // setCacheDefinitionByAlias
-        $this->util->invoke($register, 'setCacheDefinitionByAlias', $alias = 'test_alias');
-
-        // Definition
-        $definition = $this->postSetCacheDefinition($container, $tag, $id, $alias);
-        // DefinitionDecorator
-        $parent     = $container->getDefinition($definition->getParent());
-
-        // assert addMethodCalls simplified
-        $actualCalls = array_merge($parent->getMethodCalls(), $definition->getMethodCalls());
-        $this->assertSame('setArrayAccess', $actualCalls[0][0]);
-        $this->assertInstanceOf('Symfony\Component\DependencyInjection\Reference', $actualCalls[0][1][0]);
-        $this->assertFalse(isset($actualCalls[1][0]));
-        $this->assertFalse(isset($actualCalls[1][1][0]));
-    }
-
-    public function testSetConfigurationDefinition()
-    {
-        list($register, $container) = $this->createRegisterMockAndContainer();
-        $configuration = new RegisterConfiguration();
-        $id            = 'register_test';
-
-        // state not registered ID
-        $this->util->invoke($register, 'setConfigurationDefinition', $id, $configuration);
-        $this->assertTrue($container->hasDefinition($id));
-        $definition = $container->getDefinition($id);
-        $this->assertFalse($definition->isPublic());
-        $this->assertSame('YahooJapan\ConfigCacheBundle\Tests\Fixtures\RegisterConfiguration', $definition->getClass());
-
-        // state already registered ID
-        $mock = $this->createConfigurationMock();
-        $this->util->invoke($register, 'setConfigurationDefinition', $id, $mock);
-        $definition = $container->getDefinition($id);
-        $this->assertSame('YahooJapan\ConfigCacheBundle\Tests\Fixtures\RegisterConfiguration', $definition->getClass());
-        $this->assertFalse(strpos('Mock_ConfigurationInterface', $definition->getClass()) === 0);
-    }
-
-    /**
      * @dataProvider validateResourcesProvider
      */
     public function testValidateResources($resources, $expected)
@@ -1230,11 +1144,6 @@ class RegisterTest extends RegisterTestCase
     protected function createExtensionMock()
     {
         return $this->util->createInterfaceMock('Symfony\Component\DependencyInjection\Extension\ExtensionInterface');
-    }
-
-    protected function createConfigurationMock()
-    {
-        return $this->util->createInterfaceMock('Symfony\Component\Config\Definition\ConfigurationInterface');
     }
 
     protected function createResourceMock()
